Fix issues with source map loading in Firefox
This fixes a few issues using emscripten-source-map.min.js to load
source map data in Firefox. Firefox by default tries to interpret any file
loaded from an XHR as an XML file, which the source map is not, producing
an error. This tells it to load as a text file instead. This allows a successful
read from a status code of 0, which is needed if reading from a non-HTTP source,
such as a local file.